fixed int to float conversion build warnings

+++ mesmer/local/application.cpp
@@ -1689,10 +1689,10 @@ void Application::run() {
 				// loading screen
 				if (m_pre_render_frame_count == 0 || m_pre_render_frame_count == 1) {
 					ImDrawList* draw_list = ImGui::GetBackgroundDrawList();
-					draw_list->AddText(m_font_large, 48.0f, ImVec2(screenWidth / 2 - 250, screenHeight / 2), IM_COL32_WHITE, "Pre-rendering, please wait...");
-					draw_list->AddText(m_font_large, 24.0f, ImVec2(screenWidth / 2 - 100, screenHeight / 2 + 50), IM_COL32_WHITE, "-May take several seconds to a minute");
-					draw_list->AddText(m_font_large, 24.0f, ImVec2(screenWidth / 2 - 100, screenHeight / 2 + 80), IM_COL32_WHITE, "-Don't do any other task while pre-rendering");
-					// turn off explore buttons
+					draw_list->AddText(m_font_large, 48.0f, ImVec2((float)screenWidth / 2 - 250, (float)screenHeight / 2), IM_COL32_WHITE, "Pre-rendering, please wait...");
+					draw_list->AddText(m_font_large, 24.0f, ImVec2((float)screenWidth / 2 - 100, (float)screenHeight / 2 + 50), IM_COL32_WHITE, "-May take several seconds to a minute");
+					draw_list->AddText(m_font_large, 24.0f, ImVec2((float)screenWidth / 2 - 100, (float)screenHeight / 2 + 80), IM_COL32_WHITE, "-Don't do any other task while pre-rendering");
+					// turn off explore button
 					show_main_buttons = false;
 					show_demo_window = false;
 					show_fractal_selection = false;
